+++ app/Repositories/Interfaces/CurrencyAccountRepositoryInterface.php
@@ -6,7 +6,7 @@ use Illuminate\Database\Eloquent\Model;
 
 interface CurrencyAccountRepositoryInterface
 {
-    public function getUserCurrencySum(int $userId, string $currency): ?int;
+    public function getUserCurrencySum(int $userId, string $currency): ?float;
 
     public function getFirstUserCurrencyAccount(int $userId, string $currency): ?Model;
 }
